Fin de la factorielle

+++ ProjetEnCours/Labo05Biblio.h
@@ -2,4 +2,4 @@
 
 double calculerMax(double valeur1, double valeur2, double valeur3);
 
-int calculerFactorielle(int nb);
+double calculerFactorielle(int nb);
